LayoutTests/http/tests/performance/performance-resource-timing-cross-origin-media-with-cors.html

<!DOCTYPE html>
<html>
<head>
<script>
    if (window.testRunner) {
        testRunner.dumpAsText()
        testRunner.waitUntilDone();
    }
</script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<video src="http://localhost:8000/media/resources/hls/playlist-with-cookie.m3u8" crossorigin>
<script>
    var entries = new Set();
    let gotMetadata = false;
    const expectedEntries = 3;
    let runTest = function() {
        for (let entry of performance.getEntriesByType('resource')) {
            if (entries.has(entry.name))
                continue;
            entries.add(entry.name);
            if (entries.size <= expectedEntries)
                debug(entry.name);
        }

        if (gotMetadata && entries.size >= expectedEntries)
            testRunner.notifyDone();
    };
    let video = document.querySelector("video");
    setInterval(runTest, 100);

    video.addEventListener("loadedmetadata", (e) => {
        gotMetadata = true;
        if (gotMetadata && entries.size >= expectedEntries)
            testRunner.notifyDone();
    });
</script>
</body>
</html>
