LayoutTests/http/tests/performance/performance-resource-timing-cross-origin-media-with-cors-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
http://127.0.0.1:8000/resources/js-test-pre.js
http://localhost:8000/media/resources/hls/playlist-with-cookie.m3u8
http://localhost:8000/media/resources/hls/sub-playlist-with-cookie.py
